refactor(net): fold the UTF-8 codecs into bytes.ts and make mount a droppable feature
Two alignment moves from the #246 review, matching the conventions the data
modules pinned:

- framework/src/bytes.ts is the one home for module-SDK byte codecs: it
  gains stringToUtf8 (lone surrogates become U+FFFD), and net-api.ts plus
  hosts/sim/net.ts drop their hand-rolled UTF-8 encoder/decoder/length
  copies for the shared spellings.
- pocket-net follows the pocket-fs feature split: `mount` (default) carries
  the pocket-mod adapter and NetSurface; `default-features = false` drops
  pocket-mod and compiles NetCore alone for firmware with its own QuickJS
  wiring. The unused direct rquickjs dependency is gone (the adapter uses
  pocket_mod::qjs).

Verified: gate 11/11, tsc clean, cargo test -p pocket-net 5/5, clippy -D
warnings clean with and without default features, cargo check --workspace
clean.
