+++ engine/crates/pocket-net/Cargo.toml
@@ -6,10 +6,17 @@ license.workspace = true
 repository.workspace = true
 description = "Transport-neutral bounded HTTP core and PocketJS net module surface"
 
+[features]
+# `mount` brings pocket-mod (and its QuickJS embedding) for the one-line
+# globalThis.net install. A device host with its own QuickJS wiring turns
+# it off (`default-features = false`) and drives NetCore directly — the
+# MCU build then never compiles an engine it doesn't use.
+default = ["mount"]
+mount = ["dep:pocket-mod", "dep:anyhow"]
+
 [dependencies]
-pocket-mod = { workspace = true }
+pocket-mod = { workspace = true, optional = true }
 pocketjs-core = { workspace = true }
-rquickjs = { workspace = true }
-anyhow = { workspace = true }
+anyhow = { workspace = true, optional = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
